fix for TIKA-3390 contributed by kamaci (#437)
Thank you @kamaci!

+++ tika-parsers/tika-parsers-classic/tika-parsers-classic-modules/tika-parser-microsoft-module/src/main/java/org/apache/tika/parser/microsoft/SummaryExtractor.java
@@ -201,7 +201,7 @@ public class SummaryExtractor {
                     metadata.set(prop, (Double) value);
                 } else if (value instanceof Integer) {
                     Property prop = Property.externalInteger(key);
-                    metadata.set(prop, ((Integer) value).intValue());
+                    metadata.set(prop, (Integer) value);
                 }
             }
         }
